+++ shared/src/main/scala/gpdviz/Api.scala
@@ -1,5 +1,7 @@
 package gpdviz
 
+import gpdviz.model.VmSensorSystem
+
 case class ClientConfig(
                          serverName: String,
                          pusher:     Option[ClientPusherConfig]
@@ -13,4 +15,5 @@ trait Api {
 
   def clientConfig(): ClientConfig
 
+  def refresh(sysid: String): Option[VmSensorSystem]
 }